+++ imxrt-hal/src/dma.rs
@@ -153,19 +153,21 @@
 #![allow(non_snake_case)] // Compatibility with RAL
 
 mod buffer;
+mod chip;
 mod element;
 mod memcpy;
 pub(crate) mod peripheral;
 mod register;
 
 pub use buffer::{Buffer, Circular, CircularError, Drain, Linear, ReadHalf, WriteHalf};
+pub use chip::DMA_CHANNEL_COUNT;
 pub use element::Element;
 pub use memcpy::Memcpy;
 pub use peripheral::{helpers::*, Peripheral};
 
 use crate::{ccm, ral};
 use core::{
-    fmt::{self, Debug, Display},
+    fmt::{self, Debug},
     mem,
 };
 pub use register::tcd::BandwidthControl;
@@ -181,7 +183,7 @@ use register::{DMARegisters, MultiplexerRegisters, Static, DMA, MULTIPLEXER};
 /// DMA channels have very little public interface. They're best used when paired with a
 /// [`Peripheral`](struct.Peripheral.html) or a [`Memcpy`](struct.Memcpy.html).
 pub struct Channel {
-    /// Our channel number, expected to be between 0 to 31
+    /// Our channel number, expected to be between 0 and (DMA_CHANNEL_COUNT - 1)
     index: usize,
     /// Reference to the DMA registers
     registers: Static<DMARegisters>,
@@ -205,7 +207,7 @@ impl Channel {
 
     /// Returns the DMA channel number
     ///
-    /// Channels are unique and numbered within the half-open range `[0, 32)`.
+    /// Channels are unique and numbered within the half-open range `[0, DMA_CHANNEL_COUNT)`.
     pub fn channel(&self) -> usize {
         self.index
     }
@@ -457,7 +459,7 @@ pub enum Error<P> {
 /// let channel_27 = dma_channels[27].take().unwrap();
 /// let channel_0 = dma_channels[0].take().unwrap();
 /// ```
-pub struct Unclocked([Option<Channel>; 32]);
+pub struct Unclocked([Option<Channel>; DMA_CHANNEL_COUNT]);
 impl Unclocked {
     pub(crate) fn new(dma: ral::dma0::Instance, mux: ral::dmamux::Instance) -> Self {
         // Explicitly dropping instances
@@ -467,17 +469,13 @@ impl Unclocked {
         drop(dma);
         drop(mux);
 
-        Unclocked([
-            None, None, None, None, None, None, None, None, None, None, None, None, None, None,
-            None, None, None, None, None, None, None, None, None, None, None, None, None, None,
-            None, None, None, None,
-        ])
+        Unclocked(chip::DMA_CHANNEL_INIT)
     }
     /// Enable the clocks for the DMA peripheral
     ///
-    /// The return is 32 channels, each being initialized as `Some(Channel)`. Users may take channels as needed.
+    /// The return is `DMA_CHANNEL_COUNT` channels, each being initialized as `Some(Channel)`. Users may take channels as needed.
     /// The index in the array maps to the DMA channel number.
-    pub fn clock(mut self, ccm: &mut ccm::Handle) -> [Option<Channel>; 32] {
+    pub fn clock(mut self, ccm: &mut ccm::Handle) -> [Option<Channel>; DMA_CHANNEL_COUNT] {
         let (ccm, _) = ccm.raw();
         ral::modify_reg!(ral::ccm, ccm, CCGR5, CG3: 0x03);
         for (idx, channel) in self.0.iter_mut().enumerate() {
@@ -509,27 +507,6 @@ impl ErrorStatus {
 impl Debug for ErrorStatus {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "DMA_ES({:#010X})", self.es)
-    }
-}
-
-impl Display for ErrorStatus {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f,
-            "DMA_ES: VLD {vld} ECX {ecx} GPE {gpe} CPE {cpe} ERRCHN {errchn} SAE {sae} SOE {soe} DAE {dae} DOE {doe} NCE {nce} SGE {sge} SBE {sbe} DBE {dbe}",
-            vld = (self.es >> 31) & 0x1,
-            ecx = (self.es >> 16) & 0x1,
-            gpe = (self.es >> 15) & 0x1,
-            cpe = (self.es >> 14) & 0x1,
-            errchn = (self.es >> 8) & 0x1F,
-            sae = (self.es >> 7) & 0x1,
-            soe = (self.es >> 6) & 0x1,
-            dae = (self.es >> 5) & 0x1,
-            doe = (self.es >> 4) & 0x1,
-            nce = (self.es >> 3) & 0x1,
-            sge = (self.es >> 2) & 0x1,
-            sbe = (self.es >> 1) & 0x1,
-            dbe = self.es & 0x1
-        )
     }
 }
